Update algorithm to match the latest Glicko-2 documentation
On 2022-03-22, the Glicko-2 algorithm was updated, with the following
line being added to the [Glicko web page][1]:

> This document has been revised on March 22, 2022, to replace a "<"
> with "<=" in item 4(b) of Step 5 (page 3).

See the [Glicko-2 document][2] for details.

Tests for this change are missing, as at the time of writing I couldn't
figure out a test that only made use of the current public interface.

  [1]: http://glicko.net/glicko.html
  [2]: http://glicko.net/glicko/glicko2.pdf